### 2024-10-24 #665 @thomaszwagerman
Pre-Submission Enquiry: butterfly
Submitting Author Name: Thomas Zwagerman
Submitting Author Github Handle: @thomaszwagerman
Repository:  https://github.com/thomaszwagerman/butterfly
Submission type: Pre-submission

---

```
Package: butterfly
Title: Verification For Continually Updating Timeseries Data
Version: 1.0.0
    person("Thomas", "Zwagerman", , "thozwa@bas.ac.uk", role = c("aut", "cre"),
Description: Verification of continually updating timeseries data where we expect new values, but want to ensure previous data remains unchanged.
    cli,
    lifecycle,
    waldo
```

This package was written to handle the verification of continually updating time-series data, where we expect new values over time, but want to ensure previous data remains unchanged. This package provides functionality that can be used as part of a data pipeline, to check and flag changes to previous data and prevent changes going unnoticed.

N/A

Researchers, data engineers, data stewards etc. We have implemented this package in an operational data pipeline, which extracts ERA5 data and performs some calculations to generate a new publishable dataset. [ERA5 can be subject to retrospective changes](https://cds.climate.copernicus.eu/datasets/reanalysis-era5-single-levels?tab=documentation), and so to prevent these changes going unnoticed and affecting our dataset, we use butterfly to verify our data has not unexpectedly changed compared to previously published versions.

To my knowledge, there is no other package that handles the comparing of **iterations** of the same data, i.e. we want to verify there are no changes in previous data, but at the same time we do not want **new** data to throw an error and stop our pipeline.

However, my concerns are 1) this is not enough 'extra' functionality to justify a separate new rOpenSci package, and 2) the overlap with the following packages:

[waldo](https://github.com/r-lib/waldo) - `butterfly` uses `waldo::compare()` in every function to provide a report on difference. There is therefore significant overlap, however `butterfly` builds on `waldo` by providing the functionality of comparing objects where we expect changes in some places (entirely new rows of data), but not in others (previously published data). `butterfly` also provides extra user feedback to provide clarity on what it is and isn’t comparing, due to the nature of comparing only “matched” rows.

[diffdf](https://github.com/gowerc/diffdf) - similar to `waldo`, but specifically for data frames, `diffdf` provides the ability to compare data frames directly. `diffdf::diffdf()`  could have been used in our case, but I prefer `waldo`’s more explicit and clear user feedback. That said, there is significant overlap in functionality: `butterfly::loupe()` and `diffdf::diffdf_has_issues()` both provide a TRUE/FALSE difference check, while `diffdf::diffdf_issue_rows()` and `butterfly::catch()` both return the rows where changes have occurred. However, it lacks the flexibility of `butterfly` to compare object where we expect some changes, but not others.

[assertr](https://github.com/tonyfischetti/assertr) - `assertr` provides assertion functionality that can be used as part of a pipeline, and test assertions on a particular dataset, but it does not offer tools for comparison. I would use them hand-in-hand but `butterfly` has a comparison purpose.

[daquiri](https://github.com/ropensci/daiquiri/) - `daquiri` provides tools to check data quality and visually inspect timeseries data. It is a quality assurance package for timeseries, but does not have a comparison/verification purpose like `butterfly`.

N/A

The implementation of this package is very simple, almost so simple that I question whether an existing solution/package already exists. A review would be much appreciated, but getting pointed in the direction of already existing solutions to this kind of data verification problem would also be very welcome!

Many thanks :sunglasses:,

Thomas
