### 2024-10-26 #665 @adamhsparks
Hi @thomaszwagerman, thank you for your inquiry.

The editorial board has a few questions for you to consider but broadly feel that this would be in-scope.

1. Is there potential to generalise the package, _e.g._, Implement a change tolerance or examine changes in _relationships_ between variables?
2. Do you think that this work would be better as a PR contribution to the other packages that you've mentioned here that offer similar functionality?
