### 2024-10-28 #665 @thomaszwagerman
Hi @adamhsparks,

 I'm glad that this would be in-scope.

Thank you for your response and these questions, both good points. To answer your questions:

**Q1:**
The aim of this package was to generalise out from our particular ERA5 use case, and I am keen to accommodate other use cases, so for me that potential is definitely there as long as it remains within the constraints of data versioning/verification.

With regards to your particular examples:

- Change tolerance: I agree this would be a useful feature to add, with accompanying warnings that can be silenced.
- Examine changes in relationships between variables: I can also the see merit of this in model parameterisation (_e.g._ how does the relationship between _x_ and _y_ change if I change parameter _z_), so yes I think this would be a valuable addition.

**Q2:**
I think it **is** possible to incorporate the (matched comparison part) of `butterfly::loupe()` in `waldo::compare()` or `diffdf::diffdf()`, through adding an argument which specifies the unique identifier that should be used to match data frames, but operate as normal when it is NULL. This would be a non-breaking change, and might be welcomed by their respective authors?

Trickier to incorporate into other package's functionality would be the additional `cli::cat_*()` feedback (which is useful in automatic logging and stopping of pipelines), as well as the different returns implemented in butterfly (TRUE/FALSE for `loupe()`, and dataframes with/without changed rows for `catch()`/`release()`). These are quite valuable for our use, and the separation of concern (data verification/versioning vs simply quality control) is also a benefit of a separate package.
