### 2024-10-28 461c639
add notes on versioning from klump et al., imfe from siddorn et al. etc
