+++ vignettes/articles/butterfly_paper.Rmd
@@ -1,19 +1,20 @@
 ---
-title: 'butterfly: An R package for the quality assurance of continually updating timeseries'
+title: 'butterfly: An R package for the quality assurance of continually updating
+  timeseries'
 tags:
-  - R
-  - quality assurance
-  - timeseries
-  - ERA5
-authors:
-  - name: Thomas Zwagerman
-    orcid: 0000-0000-0000-0000
-    equal-contrib: true
-    affiliation: 1 
+- R
+- quality assurance
+- timeseries
+- ERA5
+date: "23 October 2024"
 affiliations:
- - name: British Antarctic Survey, UK
-   index: 1
-date: 23 October 2024
+- name: British Antarctic Survey, UK
+  index: 1
+authors:
+- name: Thomas Zwagerman
+  orcid: "0000-0000-0000-0000"
+  equal-contrib: true
+  affiliation: 1
 ---
 
 ```{r, include = FALSE}
@@ -26,6 +27,26 @@ knitr::opts_chunk$set(
 #> left out bibliography: paper.bib from yaml
 
 # Summary
+
+Importance of citing exact extract of data [(Klump et al. 2021)](https://datascience.codata.org/articles/10.5334/dsj-2021-012)
+
+Semantic versioning is widely adopted in research software [(Preston-Werner 2013)](https://semver.org/spec/v2.0.0.html)
+
+Generating a derived data product
+
+But what if you are not aware of upstream changes to your input data?
+
+A key recommendation in Siddorn et al.'s (2022) report "An Information Management Framework for Environmental Digital Twins (IMFe)...
+
+data provenance must be maintained
+
+data quality frameworks
+
+clearly documented for users and available in machine-readable format
+
+tools and methods 
+
+... for a FAIR implementation (Wilkinson et al. 2016).
 
 # Statement of Need
 At the British Antarctic Survey (BAS), we developed this package to deal with a very specific issue.
@@ -43,9 +64,9 @@ This package was originally developed to deal with [ERA5](https://cds.climate.co
 
 Usually ERA5 and ERA5T are identical, but occasionally an issue with input data can (for example for [09/21 - 12/21](https://confluence.ecmwf.int/display/CKB/ERA5T+issue+in+snow+depth), and [07/24](https://forum.ecmwf.int/t/final-validated-era5-product-to-differ-from-era5t-in-july-2024/6685)) force a recalculation, meaning previously published data differs from the final product.
 
-In most cases, this is not an issue. For static data publications which are a snapshot in time, such as data associated with a specific paper, as in "Forecasts, neural networks, and results from the paper: 'Seasonal Arctic sea ice forecasting with probabilistic deep learning'" [Andersson & Hosking (2021)](https://data.bas.ac.uk/full-record.php?id=GB/NERC/BAS/PDC/01526) or time period as in "Downscaled ERA5 monthly precipitation data using Multi-Fidelity Gaussian Processes between 1980 and 2012 for the Upper Beas and Sutlej Basins, Himalaya" [Tazi (2023)](https://data.bas.ac.uk/full-record.php?id=GB/NERC/BAS/PDC/01769), this is not an issue. These datasets clearly describe a version and time period of ERA5 from which the data were derived, and will not be amended or updated in the future, even if ERA5 is recalculated.
[…]
