+++ vignettes/articles/butterfly_paper.Rmd
@@ -12,7 +12,7 @@ affiliations:
   index: 1
 authors:
 - name: Thomas Zwagerman
-  orcid: "0000-0000-0000-0000"
+  orcid: "0009-0003-3742-3234"
   equal-contrib: true
   affiliation: 1
 ---
