### 2024-10-29 #665 @adamhsparks
@thomaszwagerman, thanks for your consideration of our suggestions and comments.

We're still happy to accept this as a submission. If you'd like to incorporate the examples from Q1 that would be welcome and would make the package more novel, but it is not totally necessary.
