+++ R/release.R
@@ -5,9 +5,7 @@
 #' which contains the new rows (if present) but matched rows which contain
 #' changes from previous data will be dropped.
 #'
-#' @param df_current data.frame, the newest/current version of dataset x.
-#' @param df_previous data.frame, the old version of dataset, for example x - t1.
-#' @param datetime_variable string, which variable to use as unique ID to join `df_current` and `df_previous`. Usually a "datetime" variable.
+#' @inheritParams create_object_list
 #' @param include_new boolean, should new rows be included? Default is TRUE.
 #'
 #' @returns A dataframe which contains only rows of `df_current` that have not changed from `df_previous`, and includes new rows.
@@ -32,7 +30,8 @@ release <- function(df_current, df_previous, datetime_variable, include_new = TR
   butterfly_object_list <- create_object_list(
     df_current,
     df_previous,
-    datetime_variable
+    datetime_variable,
+    ...
   )
 
   # By using an inner join, we drop any row which does not match in
