### 2024-10-29 638838f
properly implement waldo compare inheritance
