### 2024-10-29 8dd3bf9
reknitted documents, adding test to check extra arguments work
