+++ man/butterfly-package.Rd
@@ -4,7 +4,7 @@
 \name{butterfly-package}
 \alias{butterfly}
 \alias{butterfly-package}
-\title{butterfly: QA/QC For Continually Updating Timeseries Data}
+\title{butterfly: Verification For Continually Updating Timeseries Data}
 \description{
 \if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}
 

[…]
