### 2024-10-29 30b0c8e
turn ?waldo::compare call into code chunks
