+++ man/release.Rd
@@ -17,7 +17,7 @@ release(df_current, df_previous, datetime_variable, include_new = TRUE, ...)
 \item{include_new}{boolean, should new rows be included? Default is TRUE.}
 
 \item{...}{Other \code{waldo::compare()} arguments can be supplied here,
-such as \code{tolerance} or \code{max_diffs}. See ?waldo::compare() for a full list.}
+such as \code{tolerance} or \code{max_diffs}. See \code{?waldo::compare()} for a full list.}
 }
 \value{
 A dataframe which contains only rows of \code{df_current} that have not changed from \code{df_previous}, and includes new rows.
