### 2024-10-29 46b9141
adding additional parameter example to vignette
