### 2024-10-29 1eaf7f3
minor tweaks to paper
