### 2024-10-29 0ee04f2
updating tests and line length to comply with pkgcheck
