+++ tests/testthat/test-create_object_list.R
@@ -44,6 +44,17 @@ test_that("correct message is fed back", {
   )
 })
 
+test_that("error when datetime is not in both objects", {
+  expect_error(
+    create_object_list(
+      butterflycount$april,
+      butterflycount$march,
+      datetime_variable = "species"
+    ),
+    "`datetime_variable` must be present in `df_current` and `df_previous`"
+  )
+})
+
 test_that("a list of three objects is returned", {
   expect_length(
     create_object_list(
