+++ tests/testthat/test-catch.R
@@ -30,3 +30,16 @@ test_that("return dataframe with changed data", {
     "data.frame"
   )
 })
+
+test_that("return message when no differences", {
+  expect_output(
+    catch(
+      butterflycount$march,
+      butterflycount$february,
+      datetime_variable = "time",
+      tolerance = 2
+    ),
+    "There are no differences, so there are no rows to return.
+      Did you specify a tolerance that exceeds number of differences?"
+  )
+})
