### 2024-10-29 dbfb220
also update data to comply with pkgchecks
