+++ man/butterflycount.Rd
@@ -7,9 +7,10 @@
 \format{
 \subsection{\code{butterflycount}}{
 
-A list with 4 dataframes (january, february, march, april) containing 3 columns, and 3 + n_month rows:
+A list with 4 dataframes (january, february, march, april) containing
+3 columns, and 3 + n_month rows:
 \describe{
-\item{time}{The date on which the imaginary count took place, in yyyy-mm-dd format}
+\item{time}{The date on which the count took place, in yyyy-mm-dd format}
 \item{count}{Number of fictional butterflies counted}
 \item{species}{Butterfly species name, only appears in april}
 ...
