### 2024-10-29 2b6a2f8
add pkgcheck badge
