### 2024-10-29 439aeb0
more comprehensive summary
