### 2024-10-31 #665 @adamhsparks
@thomaszwagerman, are you OK for me to close this issue? Have we addressed all your questions sufficiently?
