### 2024-10-31 #665 @thomaszwagerman
@adamhsparks yes - many thanks for the guidance and input.

I've implemented the tolerance feature, but I need to think about how to implement checking relationships between variables. I will start a submission when I have given that some thought, and it is more appropriate to do so.
